Implementation of interactive mode of app
This is an implementation of a console application for solving quadratic equations.
Here the "readline" library is used, to read the data entered by the user, in the
code it is used in the "question" function. The user is given the opportunity to enter
the coefficients of the equation until he enters the correct data (while loop) and after
that the validated data is passed to the "quadraticEquation" function, which returns the result as a string.