Implementing of non-interactive mode of app
First, the program reads the data recorded in a text
file(coefficients separated by spaces) using library "fs".
After that  the input data is checked for validity. Then,
if the data passed the test, it is sent to the "quadraticEquation"
function, which returns the roots of the equation as a string.